Add module for all the functions
